Visual diff for Notebook files
This change introduces the feature to be able to diff ipynb from various points inside the Git experience. The currently supported points are the Changed, Staged areas, and the History panel.

The <Diff /> component maintains a registry of DiffProvider components and delegates to each provider based on the file extension.

Note
* This change is off the pre 1.0 version of the Git extension because NBDime isn't upgraded to 1.0